Separate settings menu from mapstudionew.cs (#402)
* Separate settings menu from mapstudionew

* Remove accidental folder inclusion

* Rename a var for clarity

* Rename var for clarity

* Fix menu open state + organize menus into functions
